Another reduction in the size of `dataset.py` (#10088)
* Another reduction in the size of `dataset.py`

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------
